tests: split `local` and `local type` tests
